+++ tybotRoute/CHANGELOG.md
@@ -6,7 +6,8 @@ available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
 ### 0.1.40
-- tiledesk-client => 0.9.4
+- tiledesk-client => 0.9.5
+- added _tdSendEmail directive
 
 ### 0.1.39 - online
 - ExtApi.sendSupportMessageExt => added support for rejectUnauthorized: false